feat: moving location services from REST to GraphQL
